Database is initialized with the structure

+++ vagrant/bootstrap.sh
@@ -3,12 +3,14 @@
 ## INSTALLING PHP5-CLI
 ## Installing php5-cli on the vagrant machine for PhpStorm configurations
 ####
-sudo apt-get install -y software-properties-common
-sudo apt-add-repository -y ppa:ondrej/php5-5.6
-sudo apt-get clean
-sudo apt-get update -q
-sudo apt-get install -y php5-cli php5-xdebug
-sudo cp /vagrant/20-xdebug.ini /etc/php5/cli/conf.d
+if [ ! -f /etc/php5/cli/conf.d/20-xdebug.ini ]; then
+    sudo apt-get install -y software-properties-common
+    sudo apt-add-repository -y ppa:ondrej/php5-5.6
+    sudo apt-get clean
+    sudo apt-get update -q
+    sudo apt-get install -y php5-cli php5-xdebug
+    sudo cp /vagrant/20-xdebug.ini /etc/php5/cli/conf.d
+fi
 
 ####
 ## CHECKING DOCKER INSTALLATION
@@ -24,7 +26,12 @@ docker login --email="antonio.hernandez@panamedia.net" --username="panamedia" --
 ####
 ## BUILDING DEVEL-DBMASTER CONTAINER
 ####
-sudo docker run --name devel-dbmaster -e MYSQL_ROOT_PASSWORD=tpl9 -e MYSQL_DATABASE=euromillions -d percona
+cd /docker/devel-dbmaster
+sudo docker build -t="panamedia/devel-dbmaster" .
+sudo docker stop devel-dbmaster
+sudo docker rm devel-dbmaster
+sudo docker run --name devel-dbmaster -e MYSQL_ROOT_PASSWORD=tpl9 -e MYSQL_DATABASE=euromillions -d panamedia/devel-dbmaster
+sudo docker exec -d devel-dbmaster /init_database.sh
 
 ####
 ## BUILDING DEVEL-WEB CONTAINER
